Basics of a "camera". Vertex Shader modification. Added glm
We now have the concept of a "camera" implemented - though not yet abstracted into a camera class. Currently hardcoding it to create a model, view, and projection matrix. Modified vertex shader to accept mvp matrices as uniforms. Added glm as a math library for performing matrix operations. Scene now renders a textured 2d wall that rotates on an axis