Began terrain generation
Lots to optimize, but VERY basic terrain generation now working. Still need to refactor the way models get loaded. We should only load the mesh once, but that's not what we're currently doing. Same with shaders I believe. Might make a 'model instance' type of deal where all the underlying model information gets shared, but the VAO just gets copied and bindings get done separately. idk. point is we've refactored a lot of stuff and this branch is good to go.